+++ react/components/Button/EstablishmentButtonAdd.tsx
@@ -1,5 +1,5 @@
 import React, { FC, useContext } from 'react'
-import { Button } from 'vtex.styleguide'
+import { Button, Alert } from 'vtex.styleguide'
 
 import AuthContext from '../../context/AuthContext'
 import EstablishmentContext from '../../context/EstablishmentContext'
@@ -9,17 +9,24 @@ const EstablishmentButtonAdd: FC = () => {
   const contextAuth = useContext(AuthContext)
 
   return (
-    <div className="mb5">
-      <span className="mr4">
-        <Button
-          variation="primary"
-          size="large"
-          onClick={context.saveConfigurations(contextAuth.auth.id)}
-        >
-          Adicionar
-        </Button>
-      </span>
-    </div>
+    <>
+      <div className="mb5">
+        <span className="mr4">
+          <Button
+            variation="primary"
+            size="large"
+            onClick={() => context.saveConfigurations(contextAuth.auth.id)}
+          >
+            Adicionar
+          </Button>
+        </span>
+      </div>
+      {context.showAlert === true ? (
+        <Alert type="success" onClose={context.handleCloseAlert}>
+          Dados salvos!
+        </Alert>
+      ) : null}
+    </>
   )
 }
